move vcpkg::Commands::{Hash,Fetch} to their own files
