+++ toolsrc/include/vcpkg/commands.h
@@ -11,7 +11,9 @@
 #include <vcpkg/commands.dependinfo.h>
 #include <vcpkg/commands.edit.h>
 #include <vcpkg/commands.env.h>
+#include <vcpkg/commands.fetch.h>
 #include <vcpkg/commands.format-manifest.h>
+#include <vcpkg/commands.hash.h>
 #include <vcpkg/commands.integrate.h>
 #include <vcpkg/commands.interface.h>
 #include <vcpkg/commands.list.h>
@@ -35,16 +37,6 @@ namespace vcpkg::Commands
     using CommandTypeA = void (*)(const VcpkgCmdArguments& args, const VcpkgPaths& paths, Triplet default_triplet);
     using CommandTypeB = void (*)(const VcpkgCmdArguments& args, const VcpkgPaths& paths);
     using CommandTypeC = void (*)(const VcpkgCmdArguments& args, Files::Filesystem& fs);
-
-    namespace Hash
-    {
-        void perform_and_exit(const VcpkgCmdArguments& args, const VcpkgPaths& paths);
-    }
-
-    namespace Fetch
-    {
-        void perform_and_exit(const VcpkgCmdArguments& args, const VcpkgPaths& paths);
-    }
 
     template<class T>
     struct PackageNameAndFunction
